emulator now runs in the audio thread and thus audio is fully synchronized with video

+++ src/gb.h
@@ -84,7 +84,7 @@ struct gb {
 
     u8 IE;
 
-    u16 div;
+    unsigned long div;
 
     bool prev_timer_inc;
     bool timer_overflow;
